requirements.txt
PyQt5>=5.15.9
qt-material>=2.14
Flask>=2.3.3
Flask-CORS>=4.0.0
aiohttp>=3.8.5
websocket-client>=1.6.3
pytest>=7.4.2
pytest-asyncio>=0.21.1
mypy>=1.5.1
black>=23.9.1
isort>=5.12.0
python-dotenv>=1.0.0 